Override wxApp::OnAssertFailure in L7Test.
We do nothing in this override.

wxASSERT_MSG would fail in LGridJsonCellRenderer::ProcessJsonValue and
in XmlHelper::ValidateXmlValue called by
LGridXmlCellRenderer::ProcessXmlValue. This happens with postgresql
only, when the corresponding database value is empty. The cell value
is set to [Invalid], and next rendering leads to a crash in debug build
only.
Why does wxASSERT_MSG fail? Why with postgresql and not with sqlite?